+++ src/Animation/Variables.js
@@ -50,11 +50,15 @@ const opacity = {
     },
     visible: {
         opacity: 1,
-        transition: { delay: 0.5, duration: 1 }
+        transition: { delay: 0, duration: 1 }
     },
     exit: {
         x: "-100vw",
-        transition: { ease: "easeInOut" }
+        transition: {
+            duration: 0.7,
+            ease: "easeInOut",
+            when: "beforeChildren",
+        }
     }
 }